LawAI/lawai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
bardapi==1.0.0
browser-cookie3==0.20.1
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.0
frozenlist==1.6.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
loguru==0.7.3
lxml==5.4.0
lz4==4.4.4
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.1
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pypdf==5.5.0
python-docx==1.1.2
python-dotenv==1.1.0
python-gemini-api==2.4.12
pywin32==310
requests==2.32.3
setuptools==80.8.0
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
win32_setctime==1.2.0
WMI==1.5.1
yarl==1.20.0
